+++ index.html
@@ -229,7 +229,6 @@
 
 </body>
 
-<script src="js/functions.js"></script>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 
 </html>
